requirements.txt
# requirements.txt
etcd3==0.12.0
flask==2.3.3
gevent==23.9.1
graphene==3.3
graphql-server==3.0.0b7
grpcio==1.64.1
protobuf==3.20.3
pyyaml==6.0.2
docker==7.1.0
psutil==5.9.0
